printf format warnings fixed by casting

+++ redis-cli.c
@@ -311,7 +311,8 @@ static int cliSendCommand(int argc, char **argv) {
         if (rc->flags & REDIS_CMD_MULTIBULK) {
             cmd = sdscatprintf(cmd,"*%d\r\n",argc);
             for (j = 0; j < argc; j++) {
-                cmd = sdscatprintf(cmd,"$%lu\r\n",sdslen(argv[j]));
+                cmd = sdscatprintf(cmd,"$%lu\r\n",
+                    (unsigned long)sdslen(argv[j]));
                 cmd = sdscatlen(cmd,argv[j],sdslen(argv[j]));
                 cmd = sdscatlen(cmd,"\r\n",2);
             }
@@ -319,7 +320,8 @@ static int cliSendCommand(int argc, char **argv) {
             for (j = 0; j < argc; j++) {
                 if (j != 0) cmd = sdscat(cmd," ");
                 if (j == argc-1 && rc->flags & REDIS_CMD_BULK) {
-                    cmd = sdscatprintf(cmd,"%lu",sdslen(argv[j]));
+                    cmd = sdscatprintf(cmd,"%lu",
+                        (unsigned long)sdslen(argv[j]));
                 } else {
                     cmd = sdscatlen(cmd,argv[j],sdslen(argv[j]));
                 }
